# frontend: improve `shardActiveSeriesMiddleware` performance when merging responses